+++ CHANGELOG.md
@@ -10,6 +10,10 @@ All notable changes to this project will be documented in this file.
  * `osm` example is now called `myapp` and it shows a small windows with an orthophotomap
    layer from <https://geoportal.gov.pl>.
 
+### Added
+
+ * New method: `Map::with_drawer`, which provides a simpler API for drawing custom stuff.
+
 ## 0.6.0
 
 ### Breaking